LayoutTests/platform/qt/fast/events/selectstart-by-double-triple-clicks-expected.txt

This test ensures selectstart is fired exactly once when selecting text by a mouse drag.

Initial state: PASS
Mouse down: PASS
Mouse up: PASS
Second mouse down: FAIL - expected 2 events but got 1 events
Second mouse up: PASS
Third mouse down: FAIL - expected 3 events but got 2 events
Third mouse up: FAIL - expected 3 events but got 2 events
Done.
